Remove unnecessary f's from stdlib cmath functions
powf() is a C function but std::pow() is the thin C++ wrapper that
delegates to pow(), powf(), or powl() dependent on argument type.
Most of these are already acting on floats, and those that aren't have
been fixed up to do so.

+++ src/gauge.cpp
@@ -50,15 +50,15 @@ void Gauge::destroy() {
 }
 
 void Gauge::update(float dt) {
-    const float value = std::fmaxf((float)m_min, std::fmin((float)m_max, (float)m_value));
-    const float needle_s = std::powf((value - m_min) / std::abs(m_max - m_min), m_gamma);
+    const float value = std::fmax((float)m_min, std::fmin((float)m_max, (float)m_value));
+    const float needle_s = std::pow((value - m_min) / std::abs(m_max - m_min), m_gamma);
     const float F =
         m_needleKs * (needle_s - m_needlePosition)
         - m_needleKd * m_needleVelocity;
 
     m_needleVelocity = std::fminf(
             m_needleMaxVelocity,
-            std::fmaxf(m_needleVelocity + F * dt, -m_needleMaxVelocity));
+            std::fmax(m_needleVelocity + F * dt, -m_needleMaxVelocity));
     m_needlePosition += m_needleVelocity * dt;
     m_needlePosition = std::fmax(0.0f, std::fmin(1.0f, m_needlePosition));
 }
@@ -78,7 +78,7 @@ void Gauge::render() {
     GeometryGenerator::Line2dParameters lineParams;
     generator->startShape();
     for (int i = 0; i <= std::abs(m_max - m_min); i += m_minorStep) {
-        const float s = std::powf((float)i / std::abs(m_max - m_min), m_gamma);
+        const float s = std::pow((float)i / std::abs(m_max - m_min), m_gamma);
         const float theta = s * m_thetaMax + (1 - s) * m_thetaMin;
 
         const float tickLength = (i % m_majorStep) == 0
@@ -141,14 +141,14 @@ void Gauge::render() {
         ringParams.outerRadius = outerRadius + pixelsToUnits(band.radial_offset);
         ringParams.innerRadius = outerRadius + pixelsToUnits(band.radial_offset - band.width);
 
-        const float s0 = std::powf((float)(band.start - m_min) / std::abs(m_max - m_min), m_gamma);
+        const float s0 = std::pow((float)(band.start - m_min) / std::abs(m_max - m_min), m_gamma);
         const float angle0 = s0 * m_thetaMax + (1 - s0) * m_thetaMin;
 
-        const float s1 = std::powf((float)(band.end - m_min) / std::abs(m_max - m_min), m_gamma);
+        const float s1 = std::pow((float)(band.end - m_min) / std::abs(m_max - m_min), m_gamma);
         const float angle1 = s1 * m_thetaMax + (1 - s1) * m_thetaMin;
 
-        ringParams.startAngle = std::fminf(angle0, angle1) + band.shorten_end;
-        ringParams.endAngle = std::fmaxf(angle0, angle1) - band.shorten_start;
+        ringParams.startAngle = std::fmin(angle0, angle1) + band.shorten_end;
+        ringParams.endAngle = std::fmax(angle0, angle1) - band.shorten_start;
 
         GeometryGenerator::GeometryIndices bandIndices;
         generator->startShape();
